+++ env_var.sh
@@ -5,10 +5,11 @@ ERPLIBRE_VERSION="1.6.0"
 EL_USER="erplibre"
 EL_HOME="/${EL_USER}"
 EL_HOME_ERPLIBRE="${EL_HOME}/erplibre"
-EL_HOME_ODOO="${EL_HOME_ERPLIBRE}/odoo"
+EL_HOME_ODOO="${EL_HOME_ERPLIBRE}/odoo$(cat ".odoo-version" | xargs)/odoo"
 EL_ODOO_VERSION=$(cat ".odoo-version" | xargs)
 EL_POETRY_VERSION=$(cat ".poetry-version" | xargs)
-EL_PYTHON_VERSION=$(cat ".python-version" | xargs)
+EL_PYTHON_ODOO_VERSION=$(cat ".python-odoo-version" | xargs)
+EL_PYTHON_ERPLIBRE_VERSION=$(cat "./conf/python-erplibre-version" | xargs)
 EL_ERPLIBRE_VERSION=$(cat ".erplibre-version" | xargs)
 # The default port where this Odoo instance will run under (provided you use the command -c in the terminal)
 # Set to true if you want to install it, false if you don't need it or have it already installed.
@@ -26,4 +27,4 @@ EL_INSTALL_NGINX="False"
 EL_WEBSITE_NAME="localhost"
 EL_GITHUB_TOKEN=""
 EL_MANIFEST_PROD="./default.xml"
-EL_MANIFEST_DEV="./manifest/default.dev.xml"
+EL_MANIFEST_DEV="./manifest/git_manifest_erplibre.xml"