+++ script/test/run_parallel_test.py
@@ -38,7 +38,7 @@ if not os.path.isfile(FILENAME_ODOO_VERSION):
 with open(".odoo-version", "r") as f:
     ODOO_VERSION = f.readline()
 
-LOG_FILE = "./.venv/make_test.log"
+LOG_FILE = "./.venv.erplibre/make_test.log"
 CONFIG_TESTCASE_JSON = f"./script/test/config_testcase.odoo{ODOO_VERSION}.json"